Moved forms and contexts to separate files
